Bug 1772455 - Prevent deeply nested thread messages from disappearing. r=tobyp
The `m_levels` arrays across the nsMsgDBView implementations were previously
defined as `nsTArray<uint8_t>`. When a thread reached a nesting depth greater
than 255, the `uint32_t` depth parameter was implicitly truncated during
insertion. This overflow caused the view builder to misinterpret the tree
hierarchy and drop the subsequent messages from the UI entirely.

This patch resolves the issue by upgrading `m_levels`, `m_prevLevels`, and
their associated local iterator variables from `uint8_t` to `uint32_t` to
match the public API signatures.

Differential Revision: https://phabricator.services.mozilla.com/D300525

--HG--
extra : amend_source : 36cca9bc0e4ccadf4f7d9b1080bdce8e15b728bf

+++ mailnews/base/src/nsMsgDBView.cpp
@@ -3043,7 +3043,7 @@ nsresult nsMsgDBView::PerformActionsOnJunkMsgs(bool msgsAreJunk) {
 void nsMsgDBView::ReverseThreads() {
   nsTArray<uint32_t> newFlagArray;
   nsTArray<nsMsgKey> newKeyArray;
-  nsTArray<uint8_t> newLevelArray;
+  nsTArray<uint32_t> newLevelArray;
 
   uint32_t viewSize = GetSize();
   uint32_t startThread = viewSize;
